Add files via upload
This project is an application enabling citizens and administrations to submit ideas to a vote. Depending on the outcome, the mayor's office may take steps to implement the proposed idea. On each of the proposals, a comment space allows users to debate users to discuss the proposed project and give their opinion. These comments are analyzed using an algorithm to determine the opinion of users debating and to gain an insight into the enthusiasm surrounding a proposal.
Project done in a group

Translated with www.DeepL.com/Translator (free version)